testes

+++ consulta-service/src/main/java/com/example/consulta_service/repository/ConsultaRepository.java
@@ -3,6 +3,6 @@ package com.example.consulta_service.repository;
 import com.example.consulta_service.model.Consulta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConsultaRepository extends JpaRepository<Consulta, Long> {
+public interface ConsultaRepository extends JpaRepository<Consulta, Integer> {
 
 }
